+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 HTML Template Changelog
 =======================
 
+0.0.3
+-----
+
+*   BC break: Add method `\VovanVE\HtmlTemplate\compile\CompilerInterface::getMeta()`.
+*   Deprecated: constant `\VovanVE\HtmlTemplate\Engine::VERSION`.
+
 0.0.1
 -----